Minor bugfix in Invasion percolation (5.5)
A minor issue was causing some invaded locations to be visited more
than once during an invasion run. Error was in the insertion of new
possible successors of a newly invaded site into the next-to-invade
queue, which was admitting duplicate entries into the list. In addition
to testing that these new entries are not invaded? we must check that
they are not already in the list